fix(ui): preserve opener-capable popup fallback
