+++ internal/ui/coordinator/content/popup_hosting.go
@@ -3,6 +3,7 @@ package content
 import (
 	"context"
 	"fmt"
+	"sync"
 
 	"github.com/bnema/dumber/internal/application/dto"
 	"github.com/bnema/dumber/internal/application/port"
@@ -52,6 +53,71 @@ func (pm *popupManager) openBrowserWindow(
 	}
 	if !req.NoJavaScriptAccess {
 		pm.storeReusableNamedPopupWithHost(parentPaneID, req.FrameName, paneID, popupWV, result.WindowID)
+	}
+	if lifecycle, ok := popupWV.(port.PopupLifecycleCapable); ok {
+		lifecycle.PrimePopupNavigation(req.TargetURI)
+	}
+	return popupWV
+}
+
+func popupSupportsOpenerBridge(wv port.WebView) bool {
+	_, ok := wv.(port.PopupOpenerCapable)
+	return ok
+}
+
+func (pm *popupManager) openNativePopup(
+	ctx context.Context,
+	hooks popupCoordinatorHooks,
+	parentPaneID entity.PaneID,
+	parentID port.WebViewID,
+	parentURIAtOpen string,
+	req port.PopupRequest,
+	decision dto.HostDecision,
+) port.WebView {
+	log := logging.FromContext(ctx)
+	normalized := buildPopupBrowsingContextRequest(req)
+	normalized.SourceHost = pm.resolveSourceHost(parentPaneID)
+	if pm.onOpenNativePopup == nil {
+		logBrowsingContextFailure(*log, normalized, decision, dto.BrowsingContextFailureHostUnavailable, nil)
+		return nil
+	}
+	popupWV, err := pm.createPopupWebView(ctx, parentID, req.TargetURI, false)
+	if err != nil {
+		logBrowsingContextFailure(*log, normalized, decision, dto.BrowsingContextFailureHostFailed, err)
+		return nil
+	}
+	cfg := pm.currentPopupConfig()
+	pm.setBrowsingContextDecision(popupWV, decision)
+	allowFallback := !normalized.AuthIntent && !decision.RequiresNativeOpener &&
+		(req.NoJavaScriptAccess || popupSupportsOpenerBridge(popupWV))
+	var abortOnce sync.Once
+	abortResult := false
+	onAbort := func(abortCtx context.Context, abortedWV port.WebView) bool {
+		abortOnce.Do(func() {
+			if !allowFallback {
+				logBrowsingContextFailure(*log, normalized, decision, dto.BrowsingContextFailureNativeArm, nil)
+				return
+			}
+			abortResult = pm.openExistingPopupInBrowserWindow(
+				abortCtx, hooks, parentPaneID, parentID, abortedWV, req, decision, true,
+			)
+		})
+		return abortResult
+	}
+	if err := pm.onOpenNativePopup(ctx, NativePopupInput{
+		ParentPaneID:               parentPaneID,
+		ParentWebViewID:            parentID,
+		ParentURIAtOpen:            parentURIAtOpen,
+		PopupWebView:               popupWV,
+		TargetURI:                  req.TargetURI,
+		Request:                    req,
+		ObserveOAuthAutoClose:      cfg != nil && cfg.OAuthAutoClose && IsOAuthURL(req.TargetURI),
+		AllowBrowserWindowFallback: allowFallback,
+		OnNativeHostAbort:          onAbort,
+	}); err != nil {
+		logBrowsingContextFailure(*log, normalized, decision, dto.BrowsingContextFailureNativeArm, err)
+		popupWV.Destroy()
+		return nil
 	}
 	if lifecycle, ok := popupWV.(port.PopupLifecycleCapable); ok {
 		lifecycle.PrimePopupNavigation(req.TargetURI)
