+++ internal/ui/coordinator/content/popup_manager.go
@@ -512,65 +512,6 @@ func (pm *popupManager) popupParentURIAtOpen(
 	return parentURIAtOpen
 }
 
-func (pm *popupManager) openNativePopup(
-	ctx context.Context,
-	hooks popupCoordinatorHooks,
-	parentPaneID entity.PaneID,
-	parentID port.WebViewID,
-	parentURIAtOpen string,
-	req port.PopupRequest,
-	decision dto.HostDecision,
-) port.WebView {
-	log := logging.FromContext(ctx)
-	normalized := buildPopupBrowsingContextRequest(req)
-	normalized.SourceHost = pm.resolveSourceHost(parentPaneID)
-	if pm.onOpenNativePopup == nil {
-		logBrowsingContextFailure(*log, normalized, decision, dto.BrowsingContextFailureHostUnavailable, nil)
-		return nil
-	}
-	popupWV, err := pm.createPopupWebView(ctx, parentID, req.TargetURI, false)
-	if err != nil {
-		logBrowsingContextFailure(*log, normalized, decision, dto.BrowsingContextFailureHostFailed, err)
-		return nil
-	}
-	cfg := pm.currentPopupConfig()
-	pm.setBrowsingContextDecision(popupWV, decision)
-	allowFallback := !normalized.AuthIntent && !decision.RequiresNativeOpener && req.NoJavaScriptAccess
-	var abortOnce sync.Once
-	abortResult := false
-	onAbort := func(abortCtx context.Context, abortedWV port.WebView) bool {
-		abortOnce.Do(func() {
-			if !allowFallback {
-				logBrowsingContextFailure(*log, normalized, decision, dto.BrowsingContextFailureNativeArm, nil)
-				return
-			}
-			abortResult = pm.openExistingPopupInBrowserWindow(
-				abortCtx, hooks, parentPaneID, parentID, abortedWV, req, decision, true,
-			)
-		})
-		return abortResult
-	}
-	if err := pm.onOpenNativePopup(ctx, NativePopupInput{
-		ParentPaneID:               parentPaneID,
-		ParentWebViewID:            parentID,
-		ParentURIAtOpen:            parentURIAtOpen,
-		PopupWebView:               popupWV,
-		TargetURI:                  req.TargetURI,
-		Request:                    req,
-		ObserveOAuthAutoClose:      cfg != nil && cfg.OAuthAutoClose && IsOAuthURL(req.TargetURI),
-		AllowBrowserWindowFallback: allowFallback,
-		OnNativeHostAbort:          onAbort,
-	}); err != nil {
-		logBrowsingContextFailure(*log, normalized, decision, dto.BrowsingContextFailureNativeArm, err)
-		popupWV.Destroy()
-		return nil
-	}
-	if lifecycle, ok := popupWV.(port.PopupLifecycleCapable); ok {
-		lifecycle.PrimePopupNavigation(req.TargetURI)
-	}
-	return popupWV
-}
-
 func (pm *popupManager) handlePopupCreate(
 	ctx context.Context,
 	hooks popupCoordinatorHooks,
